backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
joblib
pydantic[email]
PyJWT
